+++ frontend-sdk/packages/eslint-plugin/__fixtures__/assay-colocation/Proof.assay.test.tsx
@@ -1,6 +1,0 @@
-import { render, renderHook } from "@testing-library/react";
-import { useProofModel } from "./Proof.viewModel";
-import { ProofView } from "./Proof.view";
-
-renderHook(() => useProofModel());
-render(<ProofView />);